@@ -3,32 +3,33 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.19",
-    "@fortawesome/free-solid-svg-icons": "^5.9.0",
-    "@fortawesome/react-fontawesome": "^0.1.4",
+    "@fortawesome/fontawesome-svg-core": "^1.2.25",
+    "@fortawesome/free-solid-svg-icons": "^5.11.2",
+    "@fortawesome/react-fontawesome": "^0.1.7",
     "auth0-js": "^9.8.2",
     "bootstrap": "^4.3.1",
     "classnames": "^2.2.6",
     "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
+    "enzyme-adapter-react-16": "^1.15.1",
     "humps": "^2.0.1",
-    "jest": "24.8.0",
-    "jest-enzyme": "^7.0.2",
+    "jest": "24.9.0",
+    "jest-enzyme": "^7.1.1",
     "jquery": "^3.3.1",
     "lodash": "^4.17.14",
     "normalizr": "^3.3.0",
-    "popper.js": "^1.14.4",
+    "popper.js": "^1.16.0",
     "prop-types": "^15.6.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2",
+    "react-highlight-words": "^0.16.0",
     "react-json-view": "^1.19.1",
-    "react-redux": "^7.0.1",
-    "react-router": "^5.0.0",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "3.1.1",
-    "react-table": "^6.10.0",
-    "react-test-renderer": "^16.6.3",
-    "reactstrap": "^8.0.1",
+    "react-redux": "^7.1.1",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "3.2.0",
+    "react-table": "^6.10.3",
+    "react-test-renderer": "^16.10.2",
+    "reactstrap": "^8.1.1",
     "redux": "^4.0.1",
     "redux-actions": "^2.6.4",
     "redux-devtools-extension": "^2.13.5",
