@@ -3,6 +3,7 @@ package org.ostelco.prime.support.resources
 import arrow.core.Either
 import arrow.core.flatMap
 import arrow.core.left
+import arrow.core.right
 import io.dropwizard.auth.Auth
 import org.ostelco.prime.apierror.ApiError
 import org.ostelco.prime.apierror.ApiErrorCode
@@ -52,22 +53,25 @@ class ProfilesResource {
      * Get the subscriber profile.
      */
     @GET
-    @Path("{id}")
+    @Path("{query}")
     @Produces(MediaType.APPLICATION_JSON)
-    fun getProfile(@Auth token: AccessTokenPrincipal?,
-                   @NotNull
-                   @PathParam("id")
-                   id: String): Response =
+    fun getCustomerList(@Auth token: AccessTokenPrincipal?,
+                        @NotNull
+                        @PathParam("query")
+                        query: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                if (!isEmail(id)) {
-                    logger.info("${token.name} Accessing profile for msisdn:$id")
-                    getProfileForMsisdn(id)
+                val trimmedQuery = query.trim()
+                        .replace("'", "")
+                        .replace("\"", "")
+                if (isMsisdn(trimmedQuery)) {
+                    logger.info("${token.name} Accessing profile for msisdn: $query")
+                    getProfileListForMsisdn(trimmedQuery)
                             .responseBuilder()
                 } else {
-                    logger.info("${token.name} Accessing profile for email:$id")
-                    getProfile(contactEmail = id)
+                    logger.info("${token.name} Accessing profile for nickname/contactEmail: $query")
+                    getCustomerList(queryString = trimmedQuery)
                             .responseBuilder()
                 }
             }.build()
@@ -76,17 +80,17 @@ class ProfilesResource {
      * Get the subscriptions for this subscriber.
      */
     @GET
-    @Path("{email}/subscriptions")
+    @Path("{id}/subscriptions")
     @Produces(MediaType.APPLICATION_JSON)
     fun getSubscriptions(@Auth token: AccessTokenPrincipal?,
                          @NotNull
-                         @PathParam("email")
-                         email: String): Response =
+                         @PathParam("id")
+                         id: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                logger.info("${token.name} Accessing subscriptions for email:$email")
-                getSubscriptions(contactEmail = email)
+                logger.info("${token.name} Accessing subscriptions for customerId: $id")
+                getSubscriptions(customerId = id)
                         .responseBuilder()
             }.build()
 
@@ -95,43 +99,51 @@ class ProfilesResource {
      * Get all the eKYC scan information for this subscriber.
      */
     @GET
-    @Path("{email}/scans")
+    @Path("{id}/scans")
     @Produces(MediaType.APPLICATION_JSON)
     fun getAllScanInformation(@Auth token: AccessTokenPrincipal?,
                               @NotNull
-                              @PathParam("email")
-                              email: String): Response =
+                              @PathParam("id")
+                              id: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                logger.info("${token.name} Accessing scan information for email:$email")
-                getAllScanInformation(contactEmail = email)
+                logger.info("${token.name} Accessing scan information for customerId: $id")
+                getAllScanInformation(customerId = id)
                         .responseBuilder()
             }.build()
 
-    private fun getAllScanInformation(contactEmail: String): Either<ApiError, Collection<ScanInformation>> {
+    private fun getAllScanInformation(customerId: String): Either<ApiError, Collection<ScanInformation>> {
         return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap {identity: Identity ->
+            storage.getIdentityForCustomerId(id = customerId).flatMap { identity: Identity ->
                 storage.getAllScanInformation(identity = identity)
             }.mapLeft {
                 NotFoundError("Failed to fetch scan information.", ApiErrorCode.FAILED_TO_FETCH_SCAN_INFORMATION, it)
             }
         } catch (e: Exception) {
-            logger.error("Failed to fetch scan information for customer with contactEmail - $contactEmail", e)
+            logger.error("Failed to fetch scan information for customer with customerId: $customerId", e)
             Either.left(NotFoundError("Failed to fetch scan information", ApiErrorCode.FAILED_TO_FETCH_SCAN_INFORMATION))
         }
     }
 
     // TODO: Reuse the one from SubscriberDAO
-    private fun getProfile(contactEmail: String): Either<ApiError, Customer> {
+    private fun getCustomerList(queryString: String): Either<ApiError, Collection<Customer>> {
         return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap {identity: Identity ->
-                storage.getCustomer(identity)
-            }.mapLeft {
+            storage.getIdentitiesFor(queryString = queryString).mapLeft {
                 NotFoundError("Failed to fetch profile.", ApiErrorCode.FAILED_TO_FETCH_CUSTOMER, it)
+            }.flatMap { identityList ->
+                val customerList = identityList.mapNotNull { identity: Identity ->
+                    storage.getCustomer(identity).fold(
+                            {
+                                logger.error("Error fetching customer for $queryString,  $it")
+                                null
+                            },
+                            { it })
+                }.distinctBy { it.id }
+                Either.right(customerList)
             }
         } catch (e: Exception) {
-            logger.error("Failed to fetch profile for customer with contactEmail - $contactEmail", e)
+            logger.error("Failed to fetch profile for customer with nickname/contactEmail - $queryString", e)
             Either.left(NotFoundError("Failed to fetch profile", ApiErrorCode.FAILED_TO_FETCH_CUSTOMER))
         }
     }
@@ -142,10 +154,18 @@ class ProfilesResource {
         return pattern.matcher(email).matches()
     }
 
-    private fun getProfileForMsisdn(msisdn: String): Either<ApiError, Customer> {
+    private fun isMsisdn(query: String): Boolean {
+        val regex = "^[0-9]+$"
+        val pattern = Pattern.compile(regex)
+        return pattern.matcher(query).matches()
+    }
+
+    private fun getProfileListForMsisdn(msisdn: String): Either<ApiError, Collection<Customer>> {
         return try {
             storage.getCustomerForMsisdn(msisdn).mapLeft {
                 NotFoundError("Failed to fetch profile.", ApiErrorCode.FAILED_TO_FETCH_CUSTOMER, it)
+            }.map {
+                listOf(it)
             }
         } catch (e: Exception) {
             logger.error("Failed to fetch profile for msisdn $msisdn", e)
@@ -154,78 +174,18 @@ class ProfilesResource {
     }
 
     // TODO: Reuse the one from SubscriberDAO
-    private fun getSubscriptions(contactEmail: String): Either<ApiError, Collection<Subscription>> {
+    private fun getSubscriptions(customerId: String): Either<ApiError, Collection<Subscription>> {
         return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap {identity: Identity ->
+            storage.getIdentityForCustomerId(id = customerId).flatMap { identity: Identity ->
                 storage.getSubscriptions(identity)
             }.mapLeft {
                 NotFoundError("Failed to get subscriptions.", ApiErrorCode.FAILED_TO_FETCH_SUBSCRIPTIONS, it)
             }
         } catch (e: Exception) {
-            logger.error("Failed to get subscriptions for customer with contactEmail - $contactEmail", e)
+            logger.error("Failed to get subscriptions for customer with customerId: - $customerId", e)
             InternalServerError("Failed to get subscriptions", ApiErrorCode.FAILED_TO_FETCH_SUBSCRIPTIONS).left()
         }
     }
-
-    /**
-     * Fetches and return all plans that a subscriber subscribes
-     * to if any.
-     */
-    @GET
-    @Path("{email}/plans")
-    @Produces("application/json")
-    fun getPlans(@PathParam("email") email: String): Response =
-            storage.getIdentityForContactEmail(contactEmail = email).flatMap { identity: Identity ->
-                storage.getPlans(identity = identity)
-            }.mapLeft {
-                ApiErrorMapper.mapStorageErrorToApiError("Failed to fetch plans",
-                        ApiErrorCode.FAILED_TO_FETCH_PLANS_FOR_SUBSCRIBER,
-                        it)
-            }
-                    .responseBuilder()
-                    .build()
-
-    /**
-     * Attaches (subscribes) a subscriber to a plan.
-     */
-    @POST
-    @Path("{email}/plans/{planId}")
-    @Produces("application/json")
-    fun attachPlan(@PathParam("email") email: String,
-                   @PathParam("planId") planId: String,
-                   @QueryParam("trial_end") trialEnd: Long): Response =
-            storage.getIdentityForContactEmail(contactEmail = email).flatMap { identity: Identity ->
-                storage.subscribeToPlan(
-                        identity = identity,
-                        planId = planId,
-                        trialEnd = trialEnd)
-            }.mapLeft {
-                ApiErrorMapper.mapStorageErrorToApiError("Failed to store subscription",
-                        ApiErrorCode.FAILED_TO_STORE_SUBSCRIPTION,
-                        it)
-            }
-                    .responseBuilder(Response.Status.CREATED)
-                    .build()
-
-    /**
-     * Removes a plan from the list subscriptions for a subscriber.
-     */
-    @DELETE
-    @Path("{email}/plans/{planId}")
-    @Produces("application/json")
-    fun detachPlan(@PathParam("email") email: String,
-                   @PathParam("planId") planId: String): Response =
-            storage.getIdentityForContactEmail(contactEmail = email).flatMap { identity: Identity ->
-                storage.unsubscribeFromPlan(
-                        identity = identity,
-                        planId = planId)
-            }.mapLeft {
-                ApiErrorMapper.mapStorageErrorToApiError("Failed to remove subscription",
-                        ApiErrorCode.FAILED_TO_REMOVE_SUBSCRIPTION,
-                        it)
-            }
-                    .responseBuilder()
-                    .build()
 }
 
 /**
@@ -240,30 +200,29 @@ class BundlesResource {
      * Get all bundles for the subscriber.
      */
     @GET
-    @Path("{email}")
+    @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    fun getBundlesByEmail(@Auth token: AccessTokenPrincipal?,
-                          @NotNull
-                          @PathParam("email")
-                          email: String): Response =
+    fun getBundlesById(@Auth token: AccessTokenPrincipal?,
+                       @NotNull
+                       @PathParam("id")
+                       id: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                logger.info("${token.name} Accessing bundles for $email")
-                getBundles(contactEmail = email)
-                        .responseBuilder()
+                logger.info("${token.name} Accessing bundles for customerId: $id")
+                getBundles(customerId = id).responseBuilder()
             }.build()
 
     // TODO: Reuse the one from SubscriberDAO
-    private fun getBundles(contactEmail: String): Either<ApiError, Collection<Bundle>> {
+    private fun getBundles(customerId: String): Either<ApiError, Collection<Bundle>> {
         return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap { identity: Identity ->
+            storage.getIdentityForCustomerId(id = customerId).flatMap { identity: Identity ->
                 storage.getBundles(identity)
             }.mapLeft {
                 NotFoundError("Failed to get bundles. ${it.message}", ApiErrorCode.FAILED_TO_FETCH_BUNDLES)
             }
         } catch (e: Exception) {
-            logger.error("Failed to get bundles for customer with contactEmail - $contactEmail", e)
+            logger.error("Failed to get bundles for customer with  customerId: $customerId", e)
             Either.left(NotFoundError("Failed to get bundles", ApiErrorCode.FAILED_TO_FETCH_BUNDLES))
         }
     }
@@ -281,30 +240,29 @@ class PurchaseResource {
      * Get all purchase history for the subscriber.
      */
     @GET
-    @Path("{email}")
+    @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    fun getPurchaseHistoryByEmail(@Auth token: AccessTokenPrincipal?,
-                                  @NotNull
-                                  @PathParam("email")
-                                  email: String): Response =
+    fun getPurchaseHistoryById(@Auth token: AccessTokenPrincipal?,
+                               @NotNull
+                               @PathParam("id")
+                               id: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                logger.info("${token.name} Accessing bundles for $email")
-                getPurchaseHistory(contactEmail = email)
-                        .responseBuilder()
+                logger.info("${token.name} Accessing bundles for customerId: $id")
+                getPurchaseHistory(customerId = id).responseBuilder()
             }.build()
 
     // TODO: Reuse the one from SubscriberDAO
-    private fun getPurchaseHistory(contactEmail: String): Either<ApiError, Collection<PurchaseRecord>> {
+    private fun getPurchaseHistory(customerId: String): Either<ApiError, Collection<PurchaseRecord>> {
         return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap { identity: Identity ->
+            storage.getIdentityForCustomerId(id = customerId).flatMap { identity: Identity ->
                 storage.getPurchaseRecords(identity)
             }.bimap(
                     { NotFoundError("Failed to get purchase history.", ApiErrorCode.FAILED_TO_FETCH_PAYMENT_HISTORY, it) },
                     { it.toList() })
         } catch (e: Exception) {
-            logger.error("Failed to get purchase history for customer with contactEmail - $contactEmail", e)
+            logger.error("Failed to get purchase history for customer with customerId - $customerId", e)
             Either.left(InternalServerError("Failed to get purchase history", ApiErrorCode.FAILED_TO_FETCH_PAYMENT_HISTORY))
         }
     }
@@ -322,33 +280,33 @@ class RefundResource {
      * Refund a specified purchase for the subscriber.
      */
     @PUT
-    @Path("{email}")
+    @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    fun refundPurchaseByEmail(@Auth token: AccessTokenPrincipal?,
-                              @NotNull
-                              @PathParam("email")
-                              email: String,
-                              @NotNull
-                              @QueryParam("purchaseRecordId")
-                              purchaseRecordId: String,
-                              @NotNull
-                              @QueryParam("reason")
-                              reason: String): Response =
+    fun refundPurchaseById(@Auth token: AccessTokenPrincipal?,
+                           @NotNull
+                           @PathParam("id")
+                           id: String,
+                           @NotNull
+                           @QueryParam("purchaseRecordId")
+                           purchaseRecordId: String,
+                           @NotNull
+                           @QueryParam("reason")
+                           reason: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                logger.info("${token.name} Refunding purchase for $email at id: $purchaseRecordId")
-                refundPurchase(email, purchaseRecordId, reason)
+                logger.info("${token.name} Refunding purchase for customerId: $id at id: $purchaseRecordId")
+                refundPurchase(id, purchaseRecordId, reason)
                         .map {
-                            logger.info(NOTIFY_OPS_MARKER, "${token.name} refunded the purchase (id:$purchaseRecordId) for $email ")
+                            logger.info(NOTIFY_OPS_MARKER, "${token.name} refunded the purchase (id:$purchaseRecordId) for customerId: $id ")
                             it
                         }
                         .responseBuilder()
             }.build()
 
-    private fun refundPurchase(contactEmail: String, purchaseRecordId: String, reason: String): Either<ApiError, ProductInfo> {
+    private fun refundPurchase(customerId: String, purchaseRecordId: String, reason: String): Either<ApiError, ProductInfo> {
         return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap { identity: Identity ->
+            storage.getIdentityForCustomerId(id = customerId).flatMap { identity: Identity ->
                 storage.refundPurchase(identity, purchaseRecordId, reason)
             }.mapLeft {
                 when (it) {
@@ -357,7 +315,7 @@ class RefundResource {
                 }
             }
         } catch (e: Exception) {
-            logger.error("Failed to refund purchase for customer with contactEmail - $contactEmail, id: $purchaseRecordId", e)
+            logger.error("Failed to refund purchase for customer with customerId - $customerId, id: $purchaseRecordId", e)
             Either.left(InternalServerError("Failed to refund purchase", ApiErrorCode.FAILED_TO_REFUND_PURCHASE))
         }
     }
@@ -375,24 +333,24 @@ class ContextResource {
      * Get context for the subscriber.
      */
     @GET
-    @Path("{email}")
+    @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    fun getContextByEmail(@Auth token: AccessTokenPrincipal?,
-                          @NotNull
-                          @PathParam("email")
-                          email: String): Response =
+    fun getContextById(@Auth token: AccessTokenPrincipal?,
+                       @NotNull
+                       @PathParam("id")
+                       id: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                logger.info("${token.name} Accessing context for $email")
-                getContext(contactEmail = email)
+                logger.info("${token.name} Accessing context for customer with id: $id")
+                getContext(customerId = id)
                         .responseBuilder()
             }.build()
 
     // TODO: Reuse the one from SubscriberDAO
-    fun getContext(contactEmail: String): Either<ApiError, Context> {
+    private fun getContext(customerId: String): Either<ApiError, Context> {
         return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap { identity: Identity ->
+            storage.getIdentityForCustomerId(id = customerId).flatMap { identity: Identity ->
                 storage.getCustomer(identity).map { customer ->
                     storage.getAllRegionDetails(identity = identity)
                             .fold(
@@ -403,30 +361,8 @@ class ContextResource {
                 NotFoundError("Failed to fetch customer.", ApiErrorCode.FAILED_TO_FETCH_CONTEXT, it)
             }
         } catch (e: Exception) {
-            logger.error("Failed to fetch context for customer with contactEmail - $contactEmail", e)
+            logger.error("Failed to fetch context for customer with id - $customerId", e)
             Either.left(NotFoundError("Failed to fetch context", ApiErrorCode.FAILED_TO_FETCH_CONTEXT))
-        }
-    }
-}
-
-open class HoustonResource {
-    private val logger by getLogger()
-
-    private val storage by lazy { getResource<AdminDataSource>() }
-    private val notifier by lazy { getResource<AppNotifier>() }
-
-    fun getCustomerId(contactEmail: String): Either<ApiError, String> {
-        return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap { identity: Identity ->
-                storage.getCustomer(identity = identity)
-            }.mapLeft {
-                NotFoundError("Did not find subscription.", ApiErrorCode.FAILED_TO_FETCH_SUBSCRIPTIONS, it)
-            }.map {
-                it.id
-            }
-        } catch (e: Exception) {
-            logger.error("Did not find subscription for email $contactEmail", e)
-            InternalServerError("Did not find subscription", ApiErrorCode.FAILED_TO_FETCH_SUBSCRIPTIONS).left()
         }
     }
 }
@@ -435,39 +371,33 @@ open class HoustonResource {
  * Resource used to handle notification related REST calls.
  */
 @Path("/support/notify")
-class NotifyResource: HoustonResource() {
+class NotifyResource {
     private val logger by getLogger()
-
-    private val storage by lazy { getResource<AdminDataSource>() }
     private val notifier by lazy { getResource<AppNotifier>() }
 
     /**
      * Sends a notification to all devices for a subscriber.
      */
     @PUT
-    @Path("{email}")
+    @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    fun sendNotificationByEmail(@Auth token: AccessTokenPrincipal?,
-                                @NotNull
-                                @PathParam("email")
-                                email: String,
-                                @NotNull
-                                @QueryParam("title")
-                                title: String,
-                                @NotNull
-                                @QueryParam("message")
-                                message: String): Response =
+    fun sendNotificationById(@Auth token: AccessTokenPrincipal?,
+                             @NotNull
+                             @PathParam("id")
+                             id: String,
+                             @NotNull
+                             @QueryParam("title")
+                             title: String,
+                             @NotNull
+                             @QueryParam("message")
+                             message: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                getCustomerId(contactEmail = email)
-                        .map { customerId ->
-                            logger.info("${token.name} Sending notification to $email customerId: $customerId")
-                            val data = mapOf("timestamp" to "${System.currentTimeMillis()}")
-                            notifier.notify(customerId, title, message, data)
-                            customerId
-                        }
-                        .responseBuilder("Message Sent")
+                logger.info("${token.name} Sending notification to customerId: $id")
+                val data = mapOf("timestamp" to "${System.currentTimeMillis()}")
+                notifier.notify(id, title, message, data)
+                id.right().responseBuilder("Message Sent")
             }.build()
 }
 
@@ -476,35 +406,28 @@ class NotifyResource: HoustonResource() {
  * Resource used to handle audit log related REST calls.
  */
 @Path("/support/auditLog")
-class AuditLogResource: HoustonResource() {
-
+class AuditLogResource {
     private val logger by getLogger()
-
-    private val storage by lazy { getResource<AdminDataSource>() }
     private val auditLogStore by lazy { getResource<AuditLogStore>() }
 
     /**
      * Fetch all audit logs for a subscriber.
      */
     @GET
-    @Path("{email}")
+    @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    fun query(@Auth token: AccessTokenPrincipal?,
-              @NotNull
-              @PathParam("email")
-              email: String): Response =
+    fun queryAuditLogs(@Auth token: AccessTokenPrincipal?,
+                       @NotNull
+                       @PathParam("id")
+                       id: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                getCustomerId(contactEmail = email)
-                        .flatMap { customerId ->
-                            logger.info("${token.name} fetching audit log of $email customerId: $customerId")
-                            auditLogStore.getCustomerActivityHistory(customerId = customerId)
-                                    .mapLeft { errorMessage ->
-                                        InternalServerError(errorMessage, FAILED_TO_FETCH_AUDIT_LOGS)
-                                    }
-                        }
-                        .responseBuilder()
+                logger.info("${token.name} fetching audit log of customerId: $id")
+                auditLogStore.getCustomerActivityHistory(customerId = id)
+                        .mapLeft { errorMessage ->
+                            InternalServerError(errorMessage, FAILED_TO_FETCH_AUDIT_LOGS)
+                        }.responseBuilder()
             }.build()
 }
 
@@ -513,7 +436,6 @@ class AuditLogResource: HoustonResource() {
  */
 @Path("/support/customer")
 class CustomerResource {
-
     private val logger by getLogger()
     private val storage by lazy { getResource<AdminDataSource>() }
 
@@ -521,30 +443,29 @@ class CustomerResource {
      * Remove customer from Prime.
      */
     @DELETE
-    @Path("{email}")
+    @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
     fun removeCustomer(@Auth token: AccessTokenPrincipal?,
-              @NotNull
-              @PathParam("email")
-              email: String): Response =
+                       @NotNull
+                       @PathParam("id")
+                       id: String): Response =
             if (token == null) {
                 Response.status(Response.Status.UNAUTHORIZED)
             } else {
-                logger.info("${token.name} Removing the customer for $email")
-                removeCustomer(contactEmail = email)
-                        .responseBuilder(Response.Status.NO_CONTENT)
+                logger.info("${token.name} Removing the customer for customerId: $id")
+                removeCustomer(customerId = id).responseBuilder(Response.Status.NO_CONTENT)
             }.build()
 
     // TODO: Reuse the one from SubscriberDAO
-    private fun removeCustomer(contactEmail: String): Either<ApiError, Unit> {
+    private fun removeCustomer(customerId: String): Either<ApiError, Unit> {
         return try {
-            storage.getIdentityForContactEmail(contactEmail = contactEmail).flatMap { identity: Identity ->
+            storage.getIdentityForCustomerId(id = customerId).flatMap { identity: Identity ->
                 storage.removeCustomer(identity)
             }.mapLeft {
                 NotFoundError("Failed to remove customer.", ApiErrorCode.FAILED_TO_REMOVE_CUSTOMER, it)
             }
         } catch (e: Exception) {
-            logger.error("Failed to fetch profile for customer with contactEmail - $contactEmail", e)
+            logger.error("Failed to fetch profile for customer with customerId - $customerId", e)
             Either.left(NotFoundError("Failed to remove customer", ApiErrorCode.FAILED_TO_REMOVE_CUSTOMER))
         }
     }
