@@ -3,8 +3,6 @@ import { alertActions } from './alert.actions';
 import { CALL_API } from '../helpers/api';
 import _ from 'lodash';
 
-import { encodeEmail } from '../helpers/utils';
-
 const NOTIFY_REQUEST = 'NOTIFY_REQUEST';
 const NOTIFY_SUCCESS = 'NOTIFY_SUCCESS';
 const NOTIFY_FAILURE = 'NOTIFY_FAILURE';
@@ -39,13 +37,13 @@ const {
   setNotificationType,
  } = actions;
 
-const putNotificationByEmail = (email, title, message) => ({
+const putNotificationById = (id, title, message) => ({
   [CALL_API]: {
     actions: [
       actions.notifyRequest,
       actions.notifySuccess,
       actions.notifyFailure],
-    endpoint: `notify/${email}`,
+    endpoint: `notify/${id}`,
     method: 'PUT',
     allowEmptyResponse: true,
     params: { message, title }
@@ -57,10 +55,11 @@ const sendNotificationToSubscriber = (title, message) => (dispatch, getState) =>
     console.log('Error reported.', error);
     dispatch(alertActions.alertError(error));
   };
-  // Get the email from the fetched user
-  const subscriberEmail = encodeEmail(_.get(getState(), 'subscriber.contactEmail'));
-  if (subscriberEmail) {
-    return dispatch(putNotificationByEmail(subscriberEmail, title, message))
+
+  // Get the id from the fetched user
+  const subscriberId = _.get(getState(), 'customer.id');
+  if (subscriberId) {
+    return dispatch(putNotificationById(subscriberId, title, message))
       .catch(handleError);
   }
 };
