@@ -3,6 +3,7 @@ import { createActions } from 'redux-actions';
 
 import { CALL_API } from '../helpers/api';
 import { alertActions } from './alert.actions';
+import { currentSubscriberActions } from './currentSubscriber.actions';
 import { encodeEmail } from '../helpers/utils';
 
 const SUBSCRIBER_BY_EMAIL_REQUEST = 'SUBSCRIBER_BY_EMAIL_REQUEST';
@@ -83,87 +84,88 @@ const fetchSubscriberById = (id) => ({
   }
 });
 
-const fetchContextByEmail = (email) => ({
+const fetchContextById = (id) => ({
   [CALL_API]: {
     actions: [
       actions.contextByEmailRequest,
       actions.contextByEmailSuccess,
       actions.contextByEmailFailure],
-    endpoint: `context/${email}`,
+    endpoint: `context/${id}`,
     method: 'GET'
   }
 });
 
-const fetchSubscriptionsByEmail = (email) => ({
+const fetchSubscriptionsById = (id) => ({
   [CALL_API]: {
     actions: [
       actions.subscriptionsRequest,
       actions.subscriptionsSuccess,
       actions.subscriptionsFailure],
-    endpoint: `profiles/${email}/subscriptions`,
+    endpoint: `profiles/${id}/subscriptions`,
     method: 'GET'
   }
 });
 
-const fetchBundlesByEmail = (email) => ({
+const fetchBundlesById = (id) => ({
   [CALL_API]: {
     actions: [
       actions.bundlesRequest,
       actions.bundlesSuccess,
       actions.bundlesFailure],
-    endpoint: `bundles/${email}`,
+    endpoint: `bundles/${id}`,
     method: 'GET'
   }
 });
 
-const fetchPaymentHistoryByEmail = (email) => ({
+const fetchPaymentHistoryById = (id) => ({
   [CALL_API]: {
     actions: [
       actions.paymentHistoryRequest,
       actions.paymentHistorySuccess,
       actions.paymentHistoryFailure],
-    endpoint: `purchases/${email}`,
+    endpoint: `purchases/${id}`,
     method: 'GET'
   }
 });
 
-const putRefundPurchaseByEmail = (email, purchaseRecordId, reason) => ({
+const putRefundPurchaseById = (id, purchaseRecordId, reason) => ({
   [CALL_API]: {
     actions: [
       actions.refundPaymentRequest,
       actions.refundPaymentSuccess,
       actions.refundPaymentFailure],
-    endpoint: `refund/${email}`,
+    endpoint: `refund/${id}`,
     method: 'PUT',
     params: { purchaseRecordId, reason }
   }
 });
 
-const fetchAuditLogsByEmail = (email) => ({
+const fetchAuditLogsById = (id) => ({
   [CALL_API]: {
     actions: [
       actions.auditLogsRequest,
       actions.auditLogsSuccess,
       actions.auditLogsFailure],
-    endpoint: `auditLog/${email}`,
+    endpoint: `auditLog/${id}`,
     method: 'GET'
   }
 });
 
-const deleteUserByEmail = (email) => ({
+const deleteUserById = (id) => ({
   [CALL_API]: {
     actions: [
       actions.deleteUserRequest,
       actions.deleteUserSuccess,
       actions.deleteUserFailure],
-    endpoint: `customer/${email}`,
+    endpoint: `customer/${id}`,
     allowEmptyResponse: true,
     method: 'DELETE'
   }
 });
-
+const { selectSubscriber, clearSubscriber } = currentSubscriberActions;
 // TODO: API based implementaion. Reference: https://github.com/reduxjs/redux/issues/1676
-const getSubscriberAndBundlesByEmail = (email) => (dispatch, getState) => {
+const getSubscriberList = (email) => (dispatch, getState) => {
+  dispatch(clearSubscriber());
   localStorage.setItem('searchedEmail', email)
 
   email = encodeEmail(email);
@@ -174,56 +176,64 @@ const getSubscriberAndBundlesByEmail = (email) => (dispatch, getState) => {
 
   return dispatch(fetchSubscriberById(email))
     .then(() => {
-      // Get the email from the fetched user
-      const subscriberEmail = encodeEmail(_.get(getState(), 'subscriber.contactEmail'));
-      if (subscriberEmail) {
-        dispatch(fetchContextByEmail(subscriberEmail)).catch(handleError);
-        dispatch(fetchAuditLogsByEmail(subscriberEmail)).catch(handleError);
-        dispatch(fetchSubscriptionsByEmail(subscriberEmail)).catch(handleError);
-        return dispatch(fetchBundlesByEmail(subscriberEmail))
-          .then(() => {
-            return dispatch(fetchPaymentHistoryByEmail(subscriberEmail));
-          })
-          .catch(handleError);
+      const subscribers = _.get(getState(), 'subscribers');
+      if (Array.isArray(subscribers) && subscribers.length === 1) {
+        dispatch(selectCurrentSubscriber(subscribers[0]));
       }
     })
     .catch(handleError);
 };
 
-const refundPurchase = (purchaseRecordId, reason) => (dispatch, getState) => {
-
+const selectCurrentSubscriber = (subscriber) => (dispatch, getState) => {
+  dispatch(selectSubscriber(subscriber));
   const handleError = (error) => {
     console.log('Error reported.', error);
     dispatch(alertActions.alertError(error));
   };
 
-  // Get the email from the fetched user
-  const subscriberEmail = encodeEmail(_.get(getState(), 'subscriber.contactEmail'));
-  if (subscriberEmail) {
-    return dispatch(putRefundPurchaseByEmail(subscriberEmail, purchaseRecordId, reason))
+  dispatch(fetchContextById(subscriber.id)).catch(handleError);
+  dispatch(fetchAuditLogsById(subscriber.id)).catch(handleError);
+  dispatch(fetchSubscriptionsById(subscriber.id)).catch(handleError);
+  return dispatch(fetchBundlesById(subscriber.id))
+    .then(() => {
+      return dispatch(fetchPaymentHistoryById(subscriber.id));
+    })
+    .catch(handleError);
+};
+
+const refundPurchase = (purchaseRecordId, reason) => (dispatch, getState) => {
+  const handleError = (error) => {
+    console.log('Error reported.', error);
+    dispatch(alertActions.alertError(error));
+  };
+
+  // Get the id from the fetched user
+  const subscriberId = _.get(getState(), 'currentSubscriber.id');
+  if (subscriberId) {
+    return dispatch(putRefundPurchaseById(subscriberId, purchaseRecordId, reason))
       .then(() => {
-        return dispatch(fetchPaymentHistoryByEmail(subscriberEmail));
+        return dispatch(fetchPaymentHistoryById(subscriberId));
       })
       .catch(handleError);
   }
 };
 const deleteUser = () => (dispatch, getState) => {
-
   const handleError = (error) => {
     console.log('Error reported.', error.message);
-    let message = "Failed to delete user (" +error.message+")"
-    dispatch(alertActions.alertError({message}));
+    let message = "Failed to delete user (" + error.message + ")"
+    dispatch(alertActions.alertError({ message }));
   };
 
-  // Get the email from the fetched user
-  const subscriberEmail = encodeEmail(_.get(getState(), 'subscriber.contactEmail'));
-  if (subscriberEmail) {
-    return dispatch(deleteUserByEmail(subscriberEmail))
+  // Get the id from the fetched user
+  const subscriberId = _.get(getState(), 'currentSubscriber.id');
+  if (subscriberId) {
+    return dispatch(deleteUserById(subscriberId))
       .catch(handleError);
   }
 };
 export const subscriberActions = {
-  getSubscriberAndBundlesByEmail,
+  getSubscriberList,
+  selectCurrentSubscriber,
   refundPurchase,
   deleteUser
 };
