@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import { Button, Form, FormGroup, Label, Input } from 'reactstrap';
+import { Button, Col, Form, FormGroup, Input, Label, Row } from 'reactstrap';
 
 import { getTextType } from '../../helpers';
 
@@ -8,7 +8,8 @@ function useFormInput(initialValue, submit) {
   const [value, setValue] = useState(initialValue);
 
   function onChange(e) {
-    setValue(e.target.value);
+    const cleanValue = e.target.value.replace(/['"]+/g,'');
+    setValue(cleanValue);
   }
 
   function onSubmit(e) {
@@ -39,15 +40,21 @@ export default function SearchForm(props) {
       <Form onSubmit={input.onSubmit}>
         <FormGroup>
           <br />
-          <Label>Search user by phone number or email</Label>
-          <Input
-            type="text"
-            value={input.value}
-            onChange={input.onChange}
-            placeholder="Enter text"
-          />
+          <Label>Search user by contact Email or phone number</Label>
+          <Row>
+            <Col>
+              <Input
+                type="text"
+                value={input.value}
+                onChange={input.onChange}
+                placeholder="Enter text"
+              />
+            </Col>
+            <Col xs={"auto"} md={"auto"}>
+              <Button color="outline-primary" type="submit">Search</Button>
+            </Col>
+          </Row>
         </FormGroup>
-        <Button color="outline-primary" type="submit">Search</Button>
       </Form>
     </div>
   );
