@@ -9,11 +9,16 @@ import org.ostelco.prime.model.PurchaseRecord
 import org.ostelco.prime.model.Region
 import org.ostelco.prime.model.ScanInformation
 import org.ostelco.prime.model.Subscription
+import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.customerRegionRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.customerToBundleRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.customerToSegmentRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.customerToSimProfileRelation
+import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.exCustomerRegionRelation
+import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.exCustomerToSimProfileRelation
+import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.exSubscriptionRelation
+import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.forPurchaseByRelation
+import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.forPurchaseOfRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.identifiesRelation
-import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.purchaseRecordRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.referredRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.scanInformationRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.simProfileRegionRelation
@@ -22,6 +27,7 @@ import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.subscriptionRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.subscriptionSimProfileRelation
 import org.ostelco.prime.storage.graph.Neo4jStoreSingleton.subscriptionToBundleRelation
 import org.ostelco.prime.storage.graph.RelationType
+import org.ostelco.prime.storage.graph.model.ExCustomer
 import org.ostelco.prime.storage.graph.model.Identity
 import org.ostelco.prime.storage.graph.model.Segment
 import org.ostelco.prime.storage.graph.model.SimProfile
@@ -115,15 +121,28 @@ data class CustomerContext(override val id: String) : EntityContext<Customer>(Cu
             fromId = id,
             toId = plan.id)
 
-    infix fun purchased(product: ProductContext) = PartialRelationExpression(
-            relationType = purchaseRecordRelation,
-            fromId = id,
-            toId = product.id)
-
     infix fun belongsToSegment(segment: SegmentContext) = RelationExpression(
             relationType = customerToSegmentRelation,
             fromId = id,
             toId = segment.id)
+}
+
+data class ExCustomerContext(override val id: String) : EntityContext<ExCustomer>(ExCustomer::class, id) {
+
+    infix fun had(simProfile: SimProfileContext) = RelationExpression(
+            relationType = exCustomerToSimProfileRelation,
+            fromId = id,
+            toId = simProfile.id)
+
+    infix fun subscribedTo(subscription: SubscriptionContext) = RelationExpression(
+            relationType = exSubscriptionRelation,
+            fromId = id,
+            toId = subscription.id)
+
+    infix fun belongedTo(region: RegionContext) = RelationExpression(
+            relationType = exCustomerRegionRelation,
+            fromId = id,
+            toId = region.id)
 }
 
 data class BundleContext(override val id: String) : EntityContext<Bundle>(Bundle::class, id)
@@ -154,6 +173,21 @@ data class ScanInfoContext(override val id: String) : EntityContext<ScanInformat
 data class PlanContext(override val id: String) : EntityContext<Plan>(Plan::class, id)
 data class ProductContext(override val id: String) : EntityContext<Product>(Product::class, id)
 data class SegmentContext(override val id: String) : EntityContext<Segment>(Segment::class, id)
+
+data class PurchaseRecordContext(override val id: String) : EntityContext<PurchaseRecord>(PurchaseRecord::class, id) {
+
+    infix fun forPurchaseBy(customer: CustomerContext) = RelationExpression(
+                    relationType = forPurchaseByRelation,
+                    fromId = id,
+                    toId = customer.id
+            )
+
+    infix fun forPurchaseOf(product: ProductContext) = RelationExpression(
+                    relationType = forPurchaseOfRelation,
+                    fromId = id,
+                    toId = product.id
+            )
+}
 
 //
 // Identity
@@ -192,6 +226,12 @@ infix fun Customer.Companion.referredBy(customer: CustomerContext) =
         )
 
 //
+// ExCustomer
+//
+
+infix fun ExCustomer.Companion.withId(id: String): ExCustomerContext = ExCustomerContext(id)
+
+//
 // Bundle
 //
 
@@ -215,6 +255,18 @@ infix fun Region.Companion.linkedToSimProfile(simProfile: SimProfileContext) =
                 fromId = simProfile.id
         )
 
+infix fun Region.Companion.linkedToCustomer(customer: CustomerContext) =
+        RelatedFromClause(
+                relationType = customerRegionRelation,
+                fromId = customer.id
+        )
+
+infix fun Region.Companion.linkedToExCustomer(exCustomer: ExCustomerContext) =
+        RelatedFromClause(
+                relationType = exCustomerRegionRelation,
+                fromId = exCustomer.id
+        )
+
 //
 // SimProfiles
 //
@@ -233,6 +285,11 @@ infix fun SimProfile.Companion.forCustomer(customer: CustomerContext) =
                 fromId = customer.id
         )
 
+infix fun SimProfile.Companion.forExCustomer(exCustomer: ExCustomerContext) =
+        RelatedFromClause(
+                relationType = exCustomerToSimProfileRelation,
+                fromId = exCustomer.id
+        )
 //
 // Subscription
 //
@@ -249,6 +306,12 @@ infix fun Subscription.Companion.subscribedBy(customer: CustomerContext) =
         RelatedFromClause(
                 relationType = subscriptionRelation,
                 fromId = customer.id
+        )
+
+infix fun Subscription.Companion.wasSubscribedBy(exCustomer: ExCustomerContext) =
+        RelatedFromClause(
+                relationType = exSubscriptionRelation,
+                fromId = exCustomer.id
         )
 
 //
@@ -281,19 +344,21 @@ infix fun Plan.Companion.forCustomer(customer: CustomerContext) =
 
 infix fun Product.Companion.withSku(id: String): ProductContext = ProductContext(id)
 
-infix fun Product.Companion.purchasedBy(customer: CustomerContext) =
-        RelatedFromClause(
-                relationType = purchaseRecordRelation,
-                fromId = customer.id
-        )
-
 //
 // Purchase Record
 //
-infix fun PurchaseRecord.Companion.forPurchasesBy(customer: CustomerContext) =
-        RelationFromClause(
-                relationType = purchaseRecordRelation,
-                fromId = customer.id
+infix fun PurchaseRecord.Companion.withId(id: String): PurchaseRecordContext = PurchaseRecordContext(id)
+
+infix fun PurchaseRecord.Companion.forPurchaseBy(customer: CustomerContext) =
+        RelatedToClause(
+                relationType = forPurchaseByRelation,
+                toId = customer.id
+        )
+
+infix fun PurchaseRecord.Companion.forPurchaseOf(product: ProductContext) =
+        RelatedToClause(
+                relationType = forPurchaseOfRelation,
+                toId = product.id
         )
 
 //
