@@ -1,63 +1,16 @@
-//usr/bin/env go run "$0" "$@"; exit "$?"
-/**
- * This program is intended to be used from the command line, and will convert an
- * output file from a sim card vendor into an input file for a HSS. The assumptions
- * necessary for this to work are:
- *
- *  * The SIM card vendor produces output files similar to the example .out file
- *     found in the same source directory as this program
- *
- *  * The HSS accepts input as a CSV file, with header line 'ICCID, IMSI, KI' and subsequent
- *    lines containing ICCID/IMSI/Ki fields, all separated by commas.
- *
- * Needless to say, the outmost care should be taken when handling Ki values and
- * this program must, as a matter of course, be considered a security risk, as
- * must all  software that touch SIM values.
- *
- * With that caveat in place, the usage of this program typically looks like
- * this:
- *
- *    ./outfile_to_hss_input_converter.go  \
- *              -input-file sample_out_file_for_testing.out
- *              -output-file-prefix  ./hss-input-for-
- *
- * (followed by cryptographically strong erasure of the .out file,
- *  encapsulation of the .csv file in strong cryptography etc., none
- *  of which are handled by this script).
- */
-
-package main
+package outfileconversion
 
 import (
 	"bufio"
 	"flag"
 	"fmt"
+	"github.com/ostelco/ostelco-core/sim-administration/sim-batch-management/loltelutils"
 	"log"
 	"os"
 	"regexp"
 	"strconv"
 	"strings"
 )
-
-///
-///   Main.
-///
-
-func main() {
-	inputFile, outputFilePrefix := parseCommandLine()
-
-	fmt.Println("inputFile = ", inputFile)
-	fmt.Println("outputFilePrefix = ", outputFilePrefix)
-
-	outRecord := ReadOutputFile(inputFile)
-	outputFile := outputFilePrefix + outRecord.outputFileName + ".csv"
-	fmt.Println("outputFile = ", outputFile)
-
-	err := WriteHssCsvFile(outputFile, outRecord.entries)
-	if err != nil {
-		log.Fatal("Couldn't close output file '", outputFilePrefix, "'.  Error = '", err, "'")
-	}
-}
 
 ///
 /// Data structures
@@ -67,16 +20,16 @@ type OutputFileRecord struct {
 	Filename          string
 	inputVariables    map[string]string
 	headerDescription map[string]string
-	entries           []SimEntry
-	// TODO: As it is today, the noOfEntries is just the number of entries,
-	//       but I may want to change that to be the declared number of entries,
-	//       and then later, dynamically, read in the individual entries
+	Entries           []SimEntry
+	// TODO: As it is today, the noOfEntries is just the number of Entries,
+	//       but I may want to change that to be the declared number of Entries,
+	//       and then later, dynamically, read in the individual Entries
 	//       in a channel that is just piped to the goroutine that writes
-	//       them to file, and fails if the number of declared entries
-	//       differs from the actual number of entries.  .... but that is
+	//       them to file, and fails if the number of declared Entries
+	//       differs from the actual number of Entries.  .... but that is
 	//       for another day.
 	noOfEntries    int
-	outputFileName string
+	OutputFileName string
 }
 
 const (
@@ -107,6 +60,23 @@ type ParserState struct {
 ///  Functions
 ///
 
+
+func ParseOutputToHssConverterCommandLine() (string, string) {
+	inputFile := flag.String("input-file",
+		"not  a valid filename",
+		"path to .out file used as input file")
+
+	outputFile := flag.String("output-file-prefix",
+		"not  a valid filename",
+		"prefix to path to .csv file used as input file, filename will be autogenerated")
+
+	//
+	// Parse input according to spec above
+	//
+	flag.Parse()
+	return *inputFile, *outputFile
+}
+
 func ParseLineIntoKeyValueMap(line string, theMap map[string]string) {
 	var splitString = strings.Split(line, ":")
 	if len(splitString) != 2 {
@@ -115,10 +85,6 @@ func ParseLineIntoKeyValueMap(line string, theMap map[string]string) {
 	key := strings.TrimSpace(splitString[0])
 	value := strings.TrimSpace(splitString[1])
 	theMap[key] = value
-}
-
-func trimSuffix(s string, suffixLen int) string {
-	return s[:len(s)-suffixLen]
 }
 
 func ReadOutputFile(filename string) OutputFileRecord {
@@ -193,10 +159,10 @@ func ReadOutputFile(filename string) OutputFileRecord {
 
 			iccidWithChecksum := rawIccid
 			if strings.HasSuffix(rawIccid, "F") {
-				iccidWithChecksum = trimSuffix(rawIccid, 1)
+				iccidWithChecksum = loltelutils.TrimSuffix(rawIccid, 1)
 			}
 
-			var iccidWithoutChecksum = trimSuffix(iccidWithChecksum, 1)
+			var iccidWithoutChecksum = loltelutils.TrimSuffix(iccidWithChecksum, 1)
 			// TODO: Enable this!! checkICCIDSyntax(iccidWithChecksum)
 			entry := SimEntry{
 				rawIccid:             rawIccid,
@@ -235,9 +201,9 @@ func ReadOutputFile(filename string) OutputFileRecord {
 		Filename:          filename,
 		inputVariables:    state.inputVariables,
 		headerDescription: state.headerDescription,
-		entries:           state.entries,
+		Entries:           state.entries,
 		noOfEntries:       declaredNoOfEntities,
-		outputFileName:    getOutputFileName(state),
+		OutputFileName:    getOutputFileName(state),
 	}
 
 	return result
@@ -295,31 +261,7 @@ func isComment(s string) bool {
 	return match
 }
 
-/// XXX Add a main function that
-//   a) Reads the output file, then produces a HSS input file from that
-//   b) Later, integrate with the prime input generator, and add a
-//      database to keep track of the workflow.
-//
-//
 
-//
-// Set up command line parsing
-//
-func parseCommandLine() (string, string) {
-	inputFile := flag.String("input-file",
-		"not  a valid filename",
-		"path to .out file used as input file")
-
-	outputFile := flag.String("output-file-prefix",
-		"not  a valid filename",
-		"prefix to path to .csv file used as input file, filename will be autogenerated")
-
-	//
-	// Parse input according to spec above
-	//
-	flag.Parse()
-	return *inputFile, *outputFile
-}
 
 // fileExists checks if a file exists and is not a directory before we
 // try using it to prevent further errors.
@@ -360,3 +302,4 @@ func WriteHssCsvFile(filename string, entries []SimEntry) error {
 	fmt.Println("Successfully written ", max, " sim card records.")
 	return f.Close()
 }
+
